+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added Volume and Microphone controls for DMPS3 control systems
+ - Added XML configuration for DMPS3 volume and microphone controls
 
 ## [13.0.0] - 2018-11-20
 ### Added